Re-follow Parker if we hit a certain Twitter bug.
Twitter has this weird bug, which apparently happens to all accounts, where an
account will randomly unfollow one of their followers.  The best way to notice
this is to follow some precise number of accounts, e.g. 100, and watch that
fall slowly over time.

If Parker is still following this bot, and this bot is not following Parker,
the bot will now periodically attempt to re-follow Parker.

 * THANKS TO Noah Swartz (@SwartzCr) for pointing out this Twitter bug to me
   some time ago.